Add query all suard enabled function
